Update README.md
Recently There has been an increasing interest and demand in traveling abroad, specifically to continue the academic pursuit, one highly sought-after way of continuing this pursuit is through scholarships.
A scholarship offers students a chance to learn without having to be concerned about fees completely or partially, scholarships have been getting released over the internet under many organizations and governments in many places and websites, which has led to the need to organize and provide these scholarships in one condensed place.
The paper is organized as follows: the first chapter discusses some background information about this topic like the scholarship application process, the second chapter talks about theoretical study of the project, the third chapter reviews related works, the fourth chapter describes the system analysis, the fifth discusses project system design, and the sixth talks about implementation phase of the project.

One of the problems was finding research papers and sources talking about the core of the project which is the Information Retrieval part that uses the crawling technology, while there were projects regarding scholarships, crawlers combined with scholarship opportunities were scarce. 
The implementation of the crawler algorithm presented several significant challenges, one of which was that most websites in use today prohibit bots from scraping their content under the reCAPTCHA service, which is owned by Google and protects websites from fraud and abuse. This made the process of gathering the information required for the next step more difficult.

Another problem was the start links and stop cases which indicates the point at which the crawler should start and stop the internet search. Because of this, it is extremely expensive to crawl the internet and scrape its content. If the crawler were to answer every user search query, the service might become overloaded very quickly, endangering the server's status because it might be banned or blacklisted for making too many requests to websites.

The most significant issue we ran into when extracting information using Natural Language Processing techniques was the lack of a good dataset for training, testing, and fine-tuning; the datasets we eventually found were either too generic or too detailed for the specific type of scholarships we were looking for, making it impossible for us to get the data we needed. When we tried to use the pre-existing techniques and pretrained models, such as NER, we ran into the same issue of not being able to fine-tune or train the model on a variety of tags and named entities that were necessary for our project to define and extract information.
Another challenge we faced was implementing a LLM due to the limited hardware and software resources available for our project.
Due to the problem of not being able to access and scrape content out of some webpages 
possible solutions were found for reCAPTCHAS and getting banned such as rerouting requests through a service provider in order to avoid getting the IP address of the server banned or blacklisted, or paying for an API service to avoid reCAPTCHAS, at the current stage of the project however, these solutions were hard to be accessed due to sanctions on Syria as well as lack of financial backing for the project.

To address the challenge of not having start links and stop conditions we built a custom crawler that uses Google Search API to obtain initial links and start the crawling algorithm. for the stop conditions, we limited the algorithm for each run to stop at 10 links max.
The algorithm will not run every time the student made a search query, but rather once after a specific period of time, a week was found to be the most reasonable and at the same time with low resource and time consumption,

We used an LLM for the information extraction portion of the project. There were several routes we could have taken to either download the LLM locally, but this required storage and computational power we did not have. The second option was to clone a repository from git hub, but this required using a docker, which was difficult to implement and frequently resulted in a lag in the server connection. Ultimately, we chose the best option that allowed us to use the LLM without having to deal with the hassle of downloading it or using a middleware. This involves making API calls to the model in order to obtain responses; this is how we were able to extract the information we needed from the text that was sent to the model's API.

After we collected the information that we wanted to extract it will be stored in the database in the correct format which can be used in the system, after which the data is compared against data already saved in the database, and in case it was already added it doesn’t get added, otherwise it’s added to the system and the information is updated.